# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MARKD Application API
  description: The authoritative FastAPI application boundary for MARKD commands and
    integrations. Supabase remains the canonical data platform.
  version: 0.1.0
paths:
...
  /webhooks/whatsapp:
    get:
      tags:
      - WhatsApp
      summary: Verify Whatsapp Webhook
      operationId: verifyWhatsAppWebhook
      parameters:
      - name: hub.mode
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Hub.Mode
      - name: hub.verify_token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Hub.Verify Token
      - name: hub.challenge
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Hub.Challenge
      responses:
        '200':
          description: Successful Response
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Webhook verification was rejected.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - WhatsApp
      summary: Receive Whatsapp Webhook
      operationId: receiveWhatsAppWebhook
      parameters:
      - name: X-Hub-Signature-256
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Hub-Signature-256
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookReceipt'
        '400':
          description: Webhook payload was invalid.
        '401':
          description: Webhook signature was invalid.
        '503':
          description: Webhook evidence could not be persisted.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WebhookReceipt:
      properties:
        received:
          type: boolean
          title: Received
        inbound_messages:
          type: integer
          title: Inbound Messages
        delivery_statuses:
          type: integer
          title: Delivery Statuses
      type: object
      required:
      - received
      - inbound_messages
      - delivery_statuses
      title: WebhookReceipt
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer